+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - release name
 - fixed setting the title breaks the scrolling of the component
+- fixed guntt full screen improvements.
+- fixed searching by bool type
+- fixed fullscreen improvements tune
 
 ## [1.83.0]